Add softfloat implementation for floating-point addition and subtraction
- Introduced f32_addx.c with functions for adding and subtracting single-precision floating-point numbers.
- Implemented softfloat_subMagsF32x and softfloat_addMagsF32x for magnitude operations.
- Updated softfloat.h to declare the new f32_addx function.
- Enhanced handling of special cases such as NaN propagation and rounding modes.

+++ examples/utils/instructions.h
@@ -43,8 +43,8 @@ static inline int divx(int a, int b) {
 }
 
 /* Floating-point approximate instructions */
-static inline double faddx(double a, double b) {
-    double result;
+static inline float faddx(float a, float b) {
+    float result;
     asm volatile (
         "faddx.s %[z], %[x], %[y]\n\t"
         : [z] "=f" (result)
